\name{CCRWvsLW-package}
\alias{CCRWvsLW-package}
\alias{CCRWvsLW}
\docType{package}
\title{Searching strategy movement models}
\description{Simulate and fit searching strategy movement models to telemetry data.}
\details{
\tabular{ll}{
Package: \tab CCRWvsLW\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-08\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Marie Auger-Methe

Maintainer: Marie Auger-Methe <marie.augermethe@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
