\name{mnllTLW}
\alias{mnllTLW}
\docType{package}
\title{
Muck.
}
\description{
Muck. 
}
\usage{
mnllTLW(SL, TA_N, SLmin, SLmax)
}
\arguments{
  \item{SL}{
    double: muck
  }
  \item{TA_N}{
    double: muck
  }
  \item{SLmin}{
    double: muck
  }
  \item{SLmax}{
    double: muck
  }
}
