\name{nllBW}
\alias{nllBW}
\docType{package}
\title{
Muck.
}
\description{
Muck. 
}
\usage{
nllBW(SL,TA_N,lambda,parF=list('SLmin'=min(SL)))
}
\arguments{
  \item{SL}{
    double: muck
  }
  \item{TA_N}{
    double: muck
  }
  \item{lambda}{
    double: muck
  }
  \item{parF}{
    double: muck
  }
}
