\name{nllCRW}
\alias{nllCRW}
\docType{package}
\title{
Muck.
}
\description{
Muck. 
}
\usage{
nllCRW(SL,TA_N,kapp,lambda=parF$lambda,parF=list('SLmin'=min(SL)))
}
\arguments{
  \item{SL}{
    double: muck
  }
  \item{TA_N}{
    double: muck
  }
  \item{kapp}{
    double: muck
  }
  \item{lambda}{
    double: muck
  }
  \item{parF}{
    double: muck
  }
}
