\name{nllLW}
\alias{nllLW}
\docType{package}
\title{
Muck.
}
\description{
Muck. 
}
\usage{
nllLW(SL,TA_N,mu,parF=list('SLmin'=min(SL)))
}
\arguments{
  \item{SL}{
    double: muck
  }
  \item{TA_N}{
    double: muck
  }
  \item{mu}{
    double: muck
  }
  \item{parF}{
    double: muck
  }
}
