\name{nllTBW}
\alias{nllTBW}
\docType{package}
\title{
Muck.
}
\description{
Muck. 
}
\usage{
nllTBW(SL,TA_N,x,parF=list('SLmin'=SLmin,'SLmax'=SLmax))
}
\arguments{
  \item{SL}{
    double: muck
  }
  \item{TA_N}{
    double: muck
  }
  \item{x}{
    double: muck
  }
  \item{parF}{
    double: muck
  }
}
