\name{nllTCRW}
\alias{nllTCRW}
\docType{package}
\title{
Muck.
}
\description{
Muck. 
}
\usage{
nllTCRW(SL, TA_N, lambda, kapp, SLmin, SLmax)
}
\arguments{
  \item{SL}{
    double: muck
  }
  \item{TA_N}{
    double: muck
  }
  \item{kapp}{
    double: muck
  }
  \item{lambda}{
    double: muck
  }
  \item{SLmin}{
    double: muck
  }
  \item{SLmax}{
    double: muck
  }
}
