\name{nllTLW}
\alias{nllTLW}
\docType{package}
\title{
Muck.
}
\description{
Muck. 
}
\usage{
nllTLW(SL,TA_N,mu,parF=list(SLmin=min(SL),SLmax=max(SL)))
}
\arguments{
  \item{SL}{
    double: muck
  }
  \item{TA_N}{
    double: muck
  }
  \item{mu}{
    double: muck
  }
  \item{parF}{
    double: muck
  }
}
