\name{toMinK}
\alias{toMinK}
\docType{package}
\title{
Part of the kapp mle to minimise numerically.
}
\description{
Even with the CDLL the kappa value cannot be minimise analytically. This function is the part that needs to minimise numerically. CHECK
}
\usage{
toMinK(k,term2)	
}
\arguments{
  \item{k}{
    double: kappa value CHECK
  }
  \item{term2}{
    double: part of the derivative CHECK
  }
}
