\name{CCRWvsLW-package}
\alias{CCRWvsLW-package}
\alias{CCRWvsLW}
\docType{package}
\title{Search strategy movement models}
\description{Simulate and fit search strategy movement models to telemetry data.}
\details{
\tabular{ll}{
Package: \tab CCRWvsLW\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-21\cr
License: \tab GPL (>= 2)\cr
}
This package allows to fit a variety of random walk models to telemetry data. In particular, it allows the comparison of the TLW to a CCRW. The package also allows to simulate some of these random walks.
}
\author{
Marie Auger-Methe

Maintainer: Marie Auger-Methe <marie.augermethe@gmail.com>
}
\references{
Please refer to Auger-Methe, M., A.E. Derocher, M.J. Plank, E.A. Codling, M.A. Lewis (2015) Differentiating the Levy walk from a composite correlated random walk. arXiv:1406.4355 found at \url{http://arxiv.org/abs/1406.4355}
}
