\name{mnllCRW}
\alias{mnllCRW}
\docType{package}
\title{
Muck.
}
\description{
Muck. 
}
\usage{
mnllCRW(SL, TA_C, TA_N, SLmin, lambda=(1/mean(SL-SLmin)))
}
\arguments{
  \item{SL}{
    double: muck
  }
  \item{TA_C}{
    double: muck
  }
  \item{TA_N}{
    double: muck
  }
  \item{SLmin}{
    double: muck
  }
  \item{lambda}{
    double: muck
  }
}
