\name{mnllTBW}
\alias{mnllTBW}
\docType{package}
\title{
Muck.
}
\description{
Muck. 
}
\usage{
mnllTBW(SL, TA_N, SLmin, SLmax, lambda=(1/mean(SL-SLmin)))
}
\arguments{
  \item{SL}{
    double: muck
  }
  \item{TA_N}{
    double: muck
  }
  \item{SLmin}{
    double: muck
  }
  \item{SLmax}{
    double: muck
  }
  \item{lambda}{
    double: muck
  }
}
