\name{optK}
\alias{optK}
\docType{package}
\title{
Minimising part of kapp mle.
}
\description{
Even with the CDLL the kappa value cannot be minimise analytically. This function minimise the part that needs to be minimised numerically. 
}
\usage{
optK(t2)	
}
\arguments{
  \item{t2}{
    double: part of the derivative CHECK
  }
}
